removed Spinner.ios
