+++ packages/palette/src/elements/Spinner/index.tsx
@@ -1,2 +1,1 @@
 export * from "./Spinner"
-export * from "./Spinner.shared"
